requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# BDBF/bdbf/commands.py: 5
# BDBF/bdbf/main.py: 1
# BDBF/build/lib/bdbf/commands.py: 4
# BDBF/build/lib/bdbf/main.py: 1
# BDBF/tests/test.py: 8
discord.py == 1.4.1

# BDBF/setup.py: 1
setuptools == 49.2.0

# BDBF/docs/conf.py: 1
sphinx-rtd-theme == 0.5.0
